data connection & bottom sheet changed

+++ components/BottomSheet/index.js
@@ -5,10 +5,49 @@ import React, {
   useRef,
   useState,
 } from "react";
-import { CupertinoPane } from "cupertino-pane";
-import Close from "../icons/Close";
+/* import { CupertinoPane } from "cupertino-pane";
+ */ import Close from "../icons/Close";
 import HeaderTitle from "../HeaderTitle";
+import Sheet from "react-modal-sheet";
 
+function BottomSheet({ children, isOpen, onClose, leftComponent, title }) {
+  return (
+    <Sheet
+      isOpen={isOpen}
+      onClose={onClose}
+      snapPoints={[0.9]}
+      initialSnap={0}
+      springConfig={{
+        stiffness: 300,
+        damping: 30,
+        mass: 1,
+      }}
+    >
+      <Sheet.Container>
+        <Sheet.Header>
+          <div
+            style={{
+              display: "flex",
+              alignItems: "center",
+              justifyContent: "space-between",
+              padding: "24px",
+            }}
+          >
+            {leftComponent || <div />}
+            <HeaderTitle>{title}</HeaderTitle>
+            <Close onClick={() => onClose()} />
+          </div>
+        </Sheet.Header>
+        <Sheet.Content>{children}</Sheet.Content>
+      </Sheet.Container>
+
+      <Sheet.Backdrop />
+    </Sheet>
+  );
+}
+
+export default BottomSheet;
+/* 
 const BottomSheet = ({
   onDismiss,
   isOpen,
@@ -16,6 +55,7 @@ const BottomSheet = ({
   title,
   leftComponent,
   className,
+  close,
 }) => {
   let myPane = useRef();
 
@@ -35,6 +75,13 @@ const BottomSheet = ({
           },
           events: {
             onDidDismiss: onDismiss,
+
+            onWillDismiss: () => {
+              document.getElementsByClassName(className)[0].style.display =
+                "none";
+              document.getElementsByClassName(className)[0].style.visibility =
+                "hidden";
+            },
           },
           fastSwipeClose: true,
           bottomClose: true,
@@ -42,16 +89,20 @@ const BottomSheet = ({
         }
       );
     }
+  }, [isOpen]);
 
-    if (isOpen) {
-      myPane.current.present({ animate: true });
-    } else {
-      myPane.current.destroy({ animate: true });
+  useEffect(() => {
+    if (myPane.current) {
+      if (isOpen) {
+        myPane.current.present({ animate: true });
+      } else {
+        myPane.current.destroy({ animate: true });
+      }
     }
   }, [isOpen]);
 
   return (
-    <div className={className} style={{ visibility: "hidden" }}>
+    <div className={className}>
       <div
         style={{
           display: "flex",
@@ -75,3 +126,4 @@ const BottomSheet = ({
 };
 
 export default BottomSheet;
+ */
